Atualizações ComandaController

+++ programa/app/src/main/java/org/model/Produto.java
@@ -2,16 +2,16 @@ package org.model;
 
 public class Produto {
 	
-	private long codigo;
+	private int codigo;
 	private String nome;
 	private Categoria categoria = new Categoria();
 	private double valorUnitario;
 	
-	public long getCodigo() {
+	public int getCodigo() {
 		return codigo;
 	}
 	
-	public void setCodigo(long codigo) {
+	public void setCodigo(int codigo) {
 		this.codigo = codigo;
 	}
